move files in proper folders
