+++ src/App.tsx
@@ -11,7 +11,7 @@ import FullPageModal from './widgets/FullPageModal'
 import { LFModalProvider } from './hooks/useModal'
 import { LernfairProvider } from './hooks/useLernfair'
 import { IssueReporter } from './IssueReporter'
-import { NotificationsProvider } from './hooks/NotificationsProvider'
+import { NotificationsProvider } from './components/NotificationsProvider'
 import { WebsocketClient } from "./components/WebsocketClient"
 import { ToastNotifications } from "./components/ToastNotifications"
 import { NotificationsData } from "./components/NotificationsData"
